Version 3.0.0
A proper and best asynchronous script. (Not Auto generated)
Proper naming for scripts and classes.
For both sync and async, just httpx is sufficient as a prerequisite.
Use telegraph.close() to properly close the httpx client.
Capable of acting as Context Manager to properly close the httpx client.
New Sync and Async function to upload files.
Simple way to import required classes.

With ❤️ By @Sasivarnasarma